app/requirements.txt
Babel==2.9.1
bleach==4.1.0
blinker==1.4
click==8.0.4
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
Flask-Admin==1.6.0
Flask-Babel==2.0.0
Flask-Login==0.5.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
mysqlclient==2.1.0
packaging==21.3
passlib==1.7.4
pyparsing==3.0.7
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.32
webencodings==0.5.1
Werkzeug==2.0.3
WTForms==3.0.1
